+++ scripts/utils/data_transformations.py
@@ -1,4 +1,7 @@
 import re
+import csv
+from io import StringIO
+from subprocess import (check_output, Popen, PIPE, CalledProcessError)
 
 # TODO Incorporate column_types into function to make more efficient.
 def convert_scientific_notation(current_row):
@@ -21,6 +24,29 @@ def convert_scientific_notation(current_row):
             converted_row[key] = value
 
     return converted_row
+
+
+def generate_rows(filepath, table, **kwargs):
+    """Reads an MS Access file
+    Args:
+        filepath (str): The mdb file path.
+        table (str): The table to load.
+        kwargs (dict): Keyword arguments that are passed to the csv reader.
+    Yields:
+        dict: A row of data whose keys are the field names.
+    """
+    pkwargs = {'stdout': PIPE, 'bufsize': 1, 'universal_newlines': True}
+
+    with Popen(['mdb-export', filepath, table, '-d |'], **pkwargs).stdout as pipe:
+        first_line = StringIO(str(pipe.readline()))
+        names = next(csv.reader(first_line, **kwargs))
+        headers = [name.rstrip() for name in names]
+
+        for line in iter(pipe.readline, b''):
+            next_line = StringIO(str(line))
+            values = next(csv.reader(next_line, **kwargs))
+            values = [value.rstrip() for value in values]
+            yield dict(zip(headers, values))
 
 
 # TODO Output the location / values of the malformed rows and fields to a log file.
